+++ xenoblade_blender/import_camdo.py
@@ -4,12 +4,12 @@ import os
 import logging
 import math
 
-from .import_root import import_armature, import_root, import_images
+from .import_root import get_image_folder, import_armature, import_root, import_images
 
 from . import xc3_model_py
 
 from bpy_extras.io_utils import ImportHelper
-from bpy.props import StringProperty
+from bpy.props import StringProperty, BoolProperty
 from mathutils import Matrix
 
 
@@ -26,16 +26,28 @@ class ImportCamdo(bpy.types.Operator, ImportHelper):
         maxlen=255,
     )
 
+    pack_images: BoolProperty(
+        name="Pack Images",
+        description="Pack all images into the Blender file. Increases memory usage and import times but makes the Blender file easier to share by not creating additional files",
+        default=True
+    )
+
+    # TODO: Only show this if packed is unchecked
+    image_folder: StringProperty(
+        name="Image Folder", description="The folder for the imported images. Defaults to the file's parent folder if not set")
+
     def execute(self, context: bpy.types.Context):
         # Log any errors from Rust.
         log_fmt = '%(levelname)s %(name)s %(filename)s:%(lineno)d %(message)s'
         logging.basicConfig(format=log_fmt, level=logging.INFO)
 
-        import_camdo(context, self.filepath)
+        image_folder = get_image_folder(self.image_folder, self.filepath)
+        import_camdo(context, self.filepath, self.pack_images, image_folder)
+
         return {'FINISHED'}
 
 
-def import_camdo(context: bpy.types.Context, path: str):
+def import_camdo(context: bpy.types.Context, path: str, pack_images: bool, image_folder: str):
     start = time.time()
 
     root = xc3_model_py.load_model_legacy(path)
@@ -46,9 +58,10 @@ def import_camdo(context: bpy.types.Context, path: str):
     start = time.time()
 
     model_name = os.path.basename(path)
-    blender_images = import_images(root)
+    blender_images = import_images(
+        root, model_name, pack_images, image_folder, flip=False)
     armature = import_armature(context, root, model_name)
-    import_root(root, blender_images, armature)
+    import_root(root, blender_images, armature, flip_uvs=False)
 
     # Convert from Y up to Z up.
     armature.matrix_world = Matrix.Rotation(math.radians(90), 4, 'X')